requirements.txt
pygame-ce>=2.4.1
# PyOpenGL==3.1.10
